chore: move dataloader utils into their own file
